# openapi.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
paths:
  /:
    get:
      summary: Home
      operationId: home__get
      parameters:
        - required: false
          schema:
            title: Page Num
            type: integer
            default: 1
          name: page_num
          in: query
        - required: false
          schema:
            title: Page Size
            type: integer
            default: 9
          name: page_size
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Home_resp'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /add:
    get:
      summary: Add Get
      operationId: add_get_add_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Add Post
      operationId: add_post_add_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        '303':
          description: See Other
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Home_resp'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /update/{item_id}:
    get:
      summary: Update Get
      operationId: update_get_update__item_id__get
      parameters:
        - required: true
          schema:
            title: Item Id
            type: integer
          name: item_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not Found
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Update Post
      operationId: update_post_update__item_id__post
      parameters:
        - required: true
          schema:
            title: Item Id
            type: integer
          name: item_id
          in: path
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        '302':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Home_resp'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /delete/{item_id}:
    post:
      summary: Delete
      operationId: delete_delete__item_id__post
      parameters:
        - required: true
          schema:
            title: Item Id
            type: integer
          name: item_id
          in: path
      responses:
        '302':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Home_resp'
        '404':
          description: Not Found
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /search:
    get:
      summary: Find Items
      operationId: find_items_search_get
      parameters:
        - required: false
          schema:
            title: Name
            type: string
          name: name
          in: query
        - required: false
          schema:
            $ref: '#/components/schemas/ItType'
          name: it_type
          in: query
        - required: false
          schema:
            title: Min Price
            anyOf:
              - type: string
              - type: integer
          name: min_price
          in: query
        - required: false
          schema:
            title: Max Price
            anyOf:
              - type: string
              - type: integer
          name: max_price
          in: query
        - required: false
          schema:
            title: Page Num
            type: integer
            default: 1
          name: page_num
          in: query
        - required: false
          schema:
            title: Page Size
            type: integer
            default: 9
          name: page_size
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search_resp'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Home_resp:
      title: Home Response
      required:
        - items
        - pagination
        - pages_amt
        - items_count
        - action
      type: object
      properties:
        items:
          title: Items
          type: array
          items:
            $ref: '#/components/schemas/Item'
        pagination:
          title: Pagination
          type: object
          properties:
            next:
              type: integer
            prev:
              type: integer
        pages_amt:
          title: Pages amount
          type: integer
        items_count:
          title: Items count
          type: integer
        action:
          title: Action
          type: string
    Search_resp:
      title: Search Response
      type: object
      allOf:
        - $ref: '#/components/schemas/Home_resp'
        - type: string
          properties:
            wanted_name:
              type: string
        - type: string
          properties:
            wanted_min_price:
              type: string
        - type: string
          properties:
            wanted_max_price:
              type: string
        - type: string
          properties:
            wanted_type:
              $ref: '#/components/schemas/ItType'
    Item:
      title: Item
      required:
        - name
        - it_type
        - price
        - photo_url
      type: object
      properties:
        name:
          title: Name
          type: string
        it_type:
          $ref: '#/components/schemas/ItType'
        price:
          title: Price
          type: integer
        photo_url:
          title: Photo
          type: string
          format: binary
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ItType:
      title: ItType
      enum:
        - SMARTPHONE
        - TV
        - LAPTOP
      description: An enumeration.
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string